% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_sts}
\alias{example_sts}
\title{Example \emph{sts} dataset}
\format{
An \emph{sts} object composed of "meta" and "data" dataframes.
}
\usage{
example_sts
}
\description{
The \code{example_sts} dataset provides a quickly loadable
version of an \emph{sts} object for practicing and code examples.

This dataset was was generated on 2021-01-08 by running:

\preformatted{
library(AirSensor)

example_sts <- example_pat
example_sts$meta$elevation <- as.numeric(NA)
example_sts$meta$locationName <- example_sts$meta$label

save(example_sts, file = "data/example_sts.rda")
}
}
\keyword{datasets}
