% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaTimeSeries.R
\docType{package}
\name{MazamaTimeSeries}
\alias{MazamaTimeSeries}
\alias{MazamaTimeSeries-package}
\title{Core functionality for environmental time series}
\description{
Utility functions for working with environmental time series data from known
locations. The compact data model is structured as a list with two dataframes. A
meta' dataframe contains spatial and measuring device metadata associated with
deployments at known locations. A 'data' dataframe contains a 'datetime' column
followed by columns of measurements associated with each "device-deployment".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MazamaScience/MazamaTimeSeries}
  \item \url{https://mazamascience.github.io/MazamaTimeSeries/}
  \item Report bugs at \url{https://github.com/MazamaScience/MazamaTimeSeries/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Callahan \email{jonathan.s.callahan@gmail.com}

Other contributors:
\itemize{
  \item Hans Martin \email{hansmrtn@gmail.com} [contributor]
  \item Eli Grosman \email{eligrosman1@gmail.com} [contributor]
  \item Roger Bivand \email{roger.bivand@nhh.no} [contributor]
  \item Sebastian Luque \email{spluque@gmail.com} [contributor]
}

}
