% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_distance.R
\name{mts_distance}
\alias{mts_distance}
\title{Calculate distances from \emph{mts} locations to a location of interest}
\usage{
mts_distance(
  mts = NULL,
  longitude = NULL,
  latitude = NULL,
  measure = c("geodesic", "haversine", "vincenty", "cheap")
)
}
\arguments{
\item{mts}{\emph{mts} object.}

\item{longitude}{Longitude of the location of interest.}

\item{latitude}{Latitude of the location of interest.}

\item{measure}{One of "haversine" "vincenty", "geodesic", or "cheap"}
}
\value{
Vector of of distances (meters) named by \code{deviceDeploymentID}.
}
\description{
This function returns the distances (meters) between \code{mts} locations
and a location of interest. These distances can be used to create a
mask identifying monitors within a certain radius of the location of interest.
}
\note{
The measure \code{"cheap"} may be used to speed things up depending on
the spatial scale being considered. Distances calculated with
\code{measure = "cheap"} will vary by a few meters compared with those
calculated using \code{measure = "geodesic"}.
}
