% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mts.R
\name{mts_distinct}
\alias{mts_distinct}
\title{Retain only distinct data records in \code{mts$data}}
\usage{
mts_distinct(mts)
}
\arguments{
\item{mts}{\emph{mts} object}
}
\value{
An \emph{mts} object where each record is associated with a unique
time.
(A list with \code{meta} and \code{data} dataframes.)
}
\description{
This function is primarily for internal use.

Two successive steps are used to guarantee that the
\code{datetime} axis contains no repeated values:

\enumerate{
\item{remove any duplicate records}
\item{guarantee that rows are in \code{datetime} order}
}
}
