% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mts.R
\name{mts_extractDataFrame}
\alias{mts_extractDataFrame}
\alias{mts_extractData}
\alias{mts_extractMeta}
\title{Extract dataframes from \emph{mts} objects}
\usage{
mts_extractData(mts)

mts_extractMeta(mts)
}
\arguments{
\item{mts}{\emph{mts} object to extract dataframe from.}
}
\value{
A dataframe from the \emph{mts} object.
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise an \emph{mts} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.

\code{mts_extractData(mts)} is equivalent to \code{mts$data}.

\code{mts_extractMeta(mts)} is equivalent to \code{mts$meta}.
}
