% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_filter.R
\name{mts_filter}
\alias{mts_filter}
\title{General purpose data filtering for \emph{mts} objects}
\usage{
mts_filter(mts, ...)
}
\arguments{
\item{mts}{\emph{mts} object.}

\item{...}{Logical predicates defined in terms of the variables in the
\code{mts$data}.}
}
\value{
A subset of the incoming \code{mts}.
}
\description{
A generalized data filter for \emph{mts} objects to
choose rows/cases where conditions are true.  Multiple conditions are
combined with \code{&} or separated by a comma. Only rows where the condition
evaluates to TRUE are kept. Rows where the condition evaluates to \code{NA}
are dropped.
}
\note{
Filtering is done on non-\code{datetime} values in \code{mts$data}.
}
\examples{

library(MazamaTimeSeries)

# Are there any times when data exceeded 150?
sapply(example_mts$data, function(x) { any(x > 150, na.rm = TRUE) })

# Filter for all times where da4cadd2d6ea5302_4686 > 150
very_unhealthy <-
  example_mts \%>\%
  mts_filter(da4cadd2d6ea5302_4686 > 150)

head(very_unhealthy$data)

}
\seealso{
\link{mts_filterDate}

\link{mts_filterDatetime}
}
