% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mts.R
\name{mts_isEmpty}
\alias{mts_isEmpty}
\title{Test for an empty \emph{mts} object}
\usage{
mts_isEmpty(mts)
}
\arguments{
\item{mts}{\emph{mts} object}
}
\value{
\code{TRUE} if no data exist in \code{mts}, \code{FALSE} otherwise.
}
\description{
Convenience function for \code{nrow(mts$data) == 0}.
This makes for more readable code in functions that need to test for this.
}
\examples{
library(MazamaTimeSeries)

mts_isEmpty(example_mts)

}
