% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_select.R
\name{mts_select}
\alias{mts_select}
\title{Subset and reorder time series within an \emph{mts} object}
\usage{
mts_select(mts = NULL, deviceDeploymentID = NULL)
}
\arguments{
\item{mts}{\emph{mts} object.}

\item{deviceDeploymentID}{timeseries unique identifier}
}
\value{
A reordered (subset) of the incoming \code{mts}.
}
\description{
This function acts similarly to \code{dplyr::select()} working on
\code{mts$data}. The returned \emph{mts} object will contain only those
time series identified by \code{deviceDeploymentID} in the order specified.

This can be helpful when using faceted plot functions based on \pkg{ggplot}
such as those found in \pkg{AirMonitorPlots}.
}
\examples{

library(MazamaTimeSeries)

# Filter for "El Monte"
El_Monte <-
  example_mts \%>\%
  mts_filterMeta(communityRegion == "El Monte")

ids <- El_Monte$meta$deviceDeploymentID
rev_ids <- rev(ids)

print(ids)
print(rev_ids)

rev_El_Monte <-
  example_mts \%>\%
  mts_select(rev_ids)

print(rev_El_Monte$meta$deviceDeploymentID)

}
\seealso{
\link{mts_filterData}

\link{mts_filterDate}

\link{mts_filterDatetime}
}
