% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaTimeSeries.R
\docType{data}
\name{requiredMetaNames}
\alias{requiredMetaNames}
\title{Required columns for the 'meta' dataframe}
\format{
A vector with 10 elements
}
\usage{
requiredMetaNames
}
\description{
The 'meta' dataframe found in \emph{sts} and \emph{mts} objects
is required to have a minimum set of information for proper functioning of
the package. The names of those columns are specified in
\code{requiredMetaNames}.
}
\details{
requiredMetaNames
}
\keyword{datasets}
