% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sts.R
\name{sts_distinct}
\alias{sts_distinct}
\title{Retain only distinct data records in \code{sts$data}}
\usage{
sts_distinct(sts)
}
\arguments{
\item{sts}{\emph{sts} object}
}
\value{
An \emph{sts} object where each record is associated with a unique
time.
(A list with \code{meta} and \code{data} dataframes.)
}
\description{
Three successive steps are used to guarantee that the
\code{datetime} axis contains no repeated values:

\enumerate{
\item{remove any duplicate records}
\item{guarantee that rows are in \code{datetime} order}
\item{average together fields for any remaining records that share the same
\code{datetime}}
}
}
