% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sts.R
\name{sts_extractDataFrame}
\alias{sts_extractDataFrame}
\alias{sts_extractData}
\alias{sts_extractMeta}
\title{Extract dataframes from \emph{sts} objects}
\usage{
sts_extractData(sts)

sts_extractMeta(sts)
}
\arguments{
\item{sts}{\emph{sts} object to extract dataframe from.}
}
\value{
A dataframe from the \emph{sts} object.
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{sts} object. These functions are designed to be useful when
manipulating data in a pipeline using \code{\%>\%}.

Below is a table showing equivalent operations for each function.

\code{sts_extractData(sts)} is equivalent to \code{sts$data}.

\code{sts_extractMeta(sts)} is equivalent to \code{sts$meta}.
}
