% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_fromCSV.R
\name{sts_fromCSV}
\alias{sts_fromCSV}
\title{Load a CSV file into a SingleTimeSeries object}
\usage{
sts_fromCSV(file = NULL, nameMapping = NULL)
}
\arguments{
\item{file}{path or URL to the CSV file to load}

\item{nameMapping}{List mapping columns in the CSV file to the required columns 
specified below}
}
\value{
A \emph{sts} object containing data and metadata from the provided CSV
}
\description{
Loads a CSV file containing metadata and data from a single 
device into a SingleTimeSeries object.

The required metadata columns in the provided CSV are as follows:
\itemize{
  \item{\code{deviceID} -- device identifier (character, non-numeric)}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{timezone} -- Olson time zone}
}

Data stored in these columns will be put in the meta dataframe of the resulting 
\emph{sts} object.

The required data columns are as follows:
\itemize{
  \item{\code{datetime} -- measurement time (UTC)}
}

These columns along with any other numeric columns in the tidy dataframe will be 
put in the data dataframe of the resulting \emph{sts} object.
}
\examples{

library(MazamaTimeSeries)

# TODO: write example for sts_fromCSV()

}
