% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_fromTidyDF.R
\name{sts_fromTidyDF}
\alias{sts_fromTidyDF}
\title{Convert a tidy dataframe into a SingleTimeSeries object}
\usage{
sts_fromTidyDF(tidyDF = NULL, nameMapping = NULL)
}
\arguments{
\item{tidyDF}{Tidy dataframe containing data and metadata.}

\item{nameMapping}{List mapping columns in tidyDF to the required columns 
specified below}
}
\value{
A \emph{sts} object containing data and metadata from the original 
tidy dataframe
}
\description{
Converts a tidy dataframe containing data and metadata from
one device into a \emph{sts} object.

The required metadata columns in the provided tidy dataframe are as follows:
\itemize{
  \item{\code{deviceID} -- device identifier (character, non-numeric)}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{timezone} -- Olson time zone}
}

Data stored in these columns will be put in the meta dataframe of the resulting 
\emph{sts} object.

The required data columns are as follows:
\itemize{
  \item{\code{datetime} -- measurement time (UTC)}
}

These columns along with any other numeric columns in the tidy dataframe will be 
put in the data dataframe of the resulting \emph{sts} object.
}
\examples{

library(MazamaTimeSeries)

# TODO: write example for sts_fromTidyDF()

}
