% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sts.R
\name{sts_isEmpty}
\alias{sts_isEmpty}
\title{Test for empty \emph{sts} object}
\usage{
sts_isEmpty(sts)
}
\arguments{
\item{sts}{\emph{sts} object}
}
\value{
\code{TRUE} if no data exist in \code{sts}, \code{FALSE} otherwise.
}
\description{
Convenience function for \code{nrow(sts$data) == 0}.
This makes for more readable code in functions that need to test for this.
}
\examples{
library(MazamaTimeSeries)

sts_isEmpty(example_sts)

}
