% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_join.R
\name{sts_join}
\alias{sts_join}
\title{Join SingleTimeSeries objects along the time axis}
\usage{
sts_join(...)
}
\arguments{
\item{...}{Any number of valid SingleTimeSeries \emph{sts} objects.}
}
\value{
A SingleTimeSeries \emph{sts} object.
}
\description{
Create a merged timeseries using of any number of \emph{sts}
objects for a single sensor. If \emph{sts} objects are non-contiguous, the
resulting \emph{sts} will have gaps.
}
\note{
An error is generated if the incoming \emph{sts} objects have
non-identical metadata.
}
\examples{
library(MazamaTimeSeries)

aug01_08 <-
  example_sts \%>\%
  sts_filterDate(20180801, 20180808)

aug15_22 <-
  example_sts \%>\%
  sts_filterDate(20180815, 20180822)

aug01_22 <- sts_join(aug01_08, aug15_22)

plot(aug01_22$data$datetime)
}
