% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_mts}
\alias{example_mts}
\title{Example \emph{mts} dataset}
\format{
An \emph{mts} object composed of "meta" and "data" dataframes.
}
\usage{
example_mts
}
\description{
The \code{example_mts} dataset provides a quickly loadable
version of an \emph{mts} object for practicing and code examples.

This dataset was was generated on 2021-10-07 by running:

\preformatted{
library(AirSensor)

communities <- c("Alhambra/Monterey Park", "El Monte")

example_mts <-
  example_sensor_scaqmd \%>\%
  sensor_filterMeta(communityRegion \%in\% communities)

# Add required "locationName"
example_mts$meta$locationName <- example_mts$meta$siteName

save(example_mts, file = "data/example_mts.rda")
}
}
\keyword{datasets}
