% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mts}
\alias{example_mts}
\title{Example MultipleTimeSeries dataset}
\format{
An S3 object composed of "meta" and "data" dataframes.
}
\usage{
example_mts
}
\description{
The \code{example_mts} dataset provides a quickly loadable
version of a \emph{mts} object for practicing and code examples.

This dataset was was generated on 2021-01-09 by running:

\preformatted{
library(AirSensor)

ids <- sample(example_sensor_scaqmd$meta$deviceDeploymentID, 10)

example_mts <-
  example_sensor_scaqmd %>%
  sensor_filterMeta(deviceDeploymentID %in% ids)

save(example_mts, file = "data/example_mts.rda")
}
}
\keyword{datasets}
