% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaTimeSeries.R
\docType{package}
\name{MazamaTimeSeries}
\alias{MazamaTimeSeries}
\alias{MazamaTimeSeries-package}
\title{Core functionality for environmental time series}
\description{
Utility functions for working with environmental time series data from known
locations. The compact data model is structured as a list with two dataframes. A
meta' dataframe contains spatial and measuring device metadata associated with
deployments at known locations. A 'data' dataframe contains a 'datetime' column
followed by columns of measurements associated with each "device-deployment".
}
