% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mts.R
\name{mts_extractDataFrame}
\alias{mts_extractDataFrame}
\alias{mts_extractData}
\alias{mts_extractMeta}
\title{Extract dataframes from \emph{mts} objects}
\usage{
mts_extractData(mts)

mts_extractMeta(mts)
}
\arguments{
\item{mts}{\emph{mts} object to extract dataframe from.}
}
\value{
A dataframe from the given \emph{mts} object
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{mts} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.

Below is a table showing equivalent operations for each function.

\tabular{ll}{
  \strong{Function} \tab \strong{Equivalent Operation}\cr
  \code{mts_extractData(mts)} \tab \code{mts$data}\cr
  \code{mts_extractMeta(mts)} \tab \code{mts$meta}
}
}
