% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_trimDate.R
\name{mts_trimDate}
\alias{mts_trimDate}
\title{Trim a \emph{mts} object to full days}
\usage{
mts_trimDate(mts = NULL, timezone = NULL)
}
\arguments{
\item{mts}{\emph{mts} object.}

\item{timezone}{Olson timezone used to interpret dates.}
}
\value{
A subset of the given \emph{mts} object.
}
\description{
Trims the date range of a \emph{mts} object to local time date
boundaries which are \emph{within} the range of data. This has the effect
of removing partial-day data records at the start and end of the timeseries
and is useful when calculating full-day statistics.

Day boundaries are calculated using the specified \code{timezone} or, if
\code{NULL}, from \code{mts$meta$timezone}.
}
\examples{
library(MazamaTimeSeries)

UTC_week <- mts_filterDate(
  example_mts,
  startdate = 20190703,
  enddate = 20190706,
  timezone = "UTC"
)

# UTC day boundaries
head(UTC_week$data)

# Trim to local time day boundaries
local_week <- mts_trimDate(UTC_week)
range(local_week$data$datetime)

}
