% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sts.R
\name{sts_distinct}
\alias{sts_distinct}
\title{Retain only distinct data records in sts$data}
\usage{
sts_distinct(sts)
}
\arguments{
\item{sts}{\emph{sts} object}
}
\value{
An \emph{sts} object with no duplicated data records.
}
\description{
Three successive steps are used to guarantee that the
\code{datetime} axis contains no repeated values:

\enumerate{
\item{remove any duplicate records}
\item{guarantee that rows are in \code{datetime} order}
\item{average together fields for any remaining records that share the same
\code{datetime}}
}
}
