% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sts.R
\name{sts_extractDataFrame}
\alias{sts_extractDataFrame}
\alias{sts_extractData}
\alias{sts_extractMeta}
\title{Extract dataframes from \emph{sts} objects}
\usage{
sts_extractData(sts)

sts_extractMeta(sts)
}
\arguments{
\item{sts}{\emph{sts} object to extract dataframe from.}
}
\value{
A dataframe from the given \emph{sts} object
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{sts} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.

Below is a table showing equivalent operations for each function.

\tabular{ll}{
  \strong{Function} \tab \strong{Equivalent Operation}\cr
  \code{sts_extractData(sts)} \tab \code{sts$data}\cr
  \code{sts_extractMeta(sts)} \tab \code{sts$meta}
}
}
